+++ src/connectors/sql/stmts/create/mod.rs
@@ -6,23 +6,23 @@ pub mod database;
 pub mod table;
 pub mod index;
 
-pub struct SQLCreateStatement { }
+pub(crate) struct SQLCreateStatement { }
 
 impl SQLCreateStatement {
 
-    pub fn database(&self, database: impl Into<String>) -> SQLCreateDatabaseStatement {
+    pub(crate) fn database(&self, database: impl Into<String>) -> SQLCreateDatabaseStatement {
         SQLCreateDatabaseStatement { database: database.into(), if_not_exists: false }
     }
 
-    pub fn table(&self, table: impl Into<String>) -> SQLCreateTableStatement {
+    pub(crate) fn table(&self, table: impl Into<String>) -> SQLCreateTableStatement {
         SQLCreateTableStatement { table: table.into(), if_not_exists: false, columns: vec![] }
     }
 
-    pub fn index(&self, index: impl Into<String>) -> SQLCreateIndexStatement {
+    pub(crate) fn index(&self, index: impl Into<String>) -> SQLCreateIndexStatement {
         SQLCreateIndexStatement { unique: false, index: index.into() }
     }
 
-    pub fn unique_index(&self, index: impl Into<String>) -> SQLCreateIndexStatement {
+    pub(crate) fn unique_index(&self, index: impl Into<String>) -> SQLCreateIndexStatement {
         SQLCreateIndexStatement { unique: true, index: index.into() }
     }
 }