+++ src/connectors/sql/stmts/drop/mod.rs
@@ -6,19 +6,19 @@ pub mod database;
 pub mod table;
 pub mod index;
 
-pub struct SQLDropStatement { }
+pub(crate) struct SQLDropStatement { }
 
 impl SQLDropStatement {
 
-    pub fn database(&self, database: impl Into<String>) -> SQLDropDatabaseStatement {
+    pub(crate) fn database(&self, database: impl Into<String>) -> SQLDropDatabaseStatement {
         SQLDropDatabaseStatement { database: database.into(), if_exists: false }
     }
 
-    pub fn table(&self, table: impl Into<String>) -> SQLDropTableStatement {
+    pub(crate) fn table(&self, table: impl Into<String>) -> SQLDropTableStatement {
         SQLDropTableStatement { table: table.into(), if_exists: false }
     }
 
-    pub fn index(&self, index: impl Into<String>) -> SQLDropIndexStatement {
+    pub(crate) fn index(&self, index: impl Into<String>) -> SQLDropIndexStatement {
         SQLDropIndexStatement { index: index.into() }
     }
 }