+++ src/connectors/sql/stmts/insert_into/mod.rs
@@ -1,24 +1,24 @@
 use crate::connectors::sql::schema::dialect::SQLDialect;
 use crate::connectors::sql::schema::value::encode::ToSQLString;
 
-pub struct SQLInsertIntoStatement<'a> {
+pub(crate) struct SQLInsertIntoStatement<'a> {
     pub(crate) table: &'a str,
     pub(crate) values: Vec<(&'a str, &'a str)>,
     pub(crate) returning: Vec<String>,
 }
 
 impl<'a> SQLInsertIntoStatement<'a> {
-    pub fn value(&mut self, pair: (&'a str, &'a str)) -> &mut Self {
+    pub(crate) fn value(&mut self, pair: (&'a str, &'a str)) -> &mut Self {
         self.values.push(pair);
         self
     }
 
-    pub fn values(&mut self, pairs: Vec<(&'a str, &'a str)>) -> &mut Self {
+    pub(crate) fn values(&mut self, pairs: Vec<(&'a str, &'a str)>) -> &mut Self {
         self.values.extend(pairs);
         self
     }
 
-    pub fn returning(&mut self, keys: &Vec<String>) -> &mut Self {
+    pub(crate) fn returning(&mut self, keys: &Vec<String>) -> &mut Self {
         self.returning = keys.clone();
         self
     }