+++ src/connectors/sql/stmts/show/mod.rs
@@ -4,14 +4,14 @@ use crate::connectors::sql::stmts::show::tables::SQLShowTablesStatement;
 pub mod tables;
 pub mod index_from;
 
-pub struct SQLShowStatement { }
+pub(crate) struct SQLShowStatement { }
 
 impl SQLShowStatement {
-    pub fn tables(&self) -> SQLShowTablesStatement {
+    pub(crate) fn tables(&self) -> SQLShowTablesStatement {
         SQLShowTablesStatement { like: None }
     }
 
-    pub fn index_from(&self, table: impl Into<String>) -> SQLShowIndexFromStatement {
+    pub(crate) fn index_from(&self, table: impl Into<String>) -> SQLShowIndexFromStatement {
         SQLShowIndexFromStatement { table: table.into() }
     }
 }