+++ src/connectors/sql/stmts/use/database.rs
@@ -1,7 +1,7 @@
 use crate::connectors::sql::schema::dialect::SQLDialect;
 use crate::connectors::sql::schema::value::encode::ToSQLString;
 
-pub struct SQLUseDatabaseStatement {
+pub(crate) struct SQLUseDatabaseStatement {
     pub(crate) database: String
 }